SWD-26 generate mapper for every dto

+++ backend/src/main/java/com/fptgang/backend/repository/CategoryRepos.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoryRepos extends JpaRepository<Category, Integer> , JpaSpecificationExecutor<Category> {
+public interface CategoryRepos extends JpaRepository<Category, Long> , JpaSpecificationExecutor<Category> {
 }
